Change "long" transformations to "int"

+++ app/src/main/java/com/gianxd/audiodev/activity/LocalStreamActivity.java
@@ -839,8 +839,8 @@ public class LocalStreamActivity extends  AppCompatActivity  {
 						miniplayerSongArtist.setText(musicData.get(Integer.parseInt(profileData.get("lastSongItemPosition").toString())).get("songArtist").toString());
 						miniplayerSeekbar.setMax(playbackSrv.getMaxDuration());
 						miniplayerSeekbar.setProgress(playbackSrv.getCurrentPosition());
-						maxDuration.setText(String.valueOf((long)((playbackSrv.getMaxDuration() / 1000) / 60)).concat(":".concat(new DecimalFormat("00").format((playbackSrv.getMaxDuration() / 1000) % 60))));
-						currentDuration.setText(String.valueOf((long)((playbackSrv.getCurrentPosition() / 1000) / 60)).concat(":".concat(new DecimalFormat("00").format((playbackSrv.getCurrentPosition() / 1000) % 60))));
+						maxDuration.setText(String.valueOf((int) ((playbackSrv.getMaxDuration() / 1000) / 60)).concat(":".concat(new DecimalFormat("00").format((playbackSrv.getMaxDuration() / 1000) % 60))));
+						currentDuration.setText(String.valueOf((int)((playbackSrv.getCurrentPosition() / 1000) / 60)).concat(":".concat(new DecimalFormat("00").format((playbackSrv.getCurrentPosition() / 1000) % 60))));
 						seekbarDuration.setMax(playbackSrv.getMaxDuration());
 						seekbarDuration.setProgress(playbackSrv.getCurrentPosition());
 					} else {
